+++ src/main/java/model/LadderResult.java
@@ -1,6 +1,6 @@
 package model;
 
-import util.ExceptionMessage;
+import constant.ExceptionMessage;
 
 import java.util.Arrays;
 import java.util.Collections;
